(#6814) - Fix and move test for support for cookies on Node.js
